fix: 채팅 페이지 통합 및 네브바 연결
